+++ tests/models/glpn_kitti/test_glpn_kitti.py
@@ -10,7 +10,7 @@ from tests.utils import ModelTester
 class ThisTester(ModelTester):
     def _load_model(self):
         self.processor = GLPNImageProcessor.from_pretrained("vinvino02/glpn-kitti")
-        model = GLPNForDepthEstimation.from_pretrained("vinvino02/glpn-kitti")
+        model = GLPNForDepthEstimation.from_pretrained("vinvino02/glpn-kitti", torch_dtype=torch.bfloat16)
         return model
 
     def _load_inputs(self):
@@ -18,6 +18,7 @@ class ThisTester(ModelTester):
         self.image = Image.open(requests.get(url, stream=True).raw)
         # prepare image for the model
         inputs = self.processor(images=self.image, return_tensors="pt")
+        inputs["pixel_values"] = inputs["pixel_values"].to(torch.bfloat16)
         return inputs
 
 
@@ -45,7 +46,7 @@ def test_glpn_kitti(record_property, mode):
         )
 
         # visualize the prediction
-        output = prediction.squeeze().cpu().numpy()
+        output = prediction.squeeze().cpu().to(float).numpy()
         formatted = (output * 255 / np.max(output)).astype("uint8")
         depth = Image.fromarray(formatted)
